Support multi-cost query in single command
This feature requires latest alto client library.

The rucio user can specify a `metrics` attribute in the `[client]`
section of the `alto.conf`. The value should be a JSON object like the
following:

```
{
  "metric1": {
    "url": "https://alto.example.com/costmap",
    "resource-type": "cost-map",
    "auth": ["username", "passw0rd"]
  },
  ...
}
```

Then the rucio client can sort a file replica using the ALTO-based
sorter as the following:

```
rucio list-file-replicas \
    --sort='alto,by=metric1;metric2,where=metric3>2' \
    --metalink \
    test:file1
```

Signed-off-by: Jensen Zhang <hack@jensen-zhang.site>